move util function to scaler.go file
